refactor(mobile): Convert mobile screens to TypeScript
