+++ lib/components/mobile/navigation-bar.js
@@ -19,7 +19,7 @@ class MobileNavigationBar extends Component {
     configLanguages: PropTypes.object,
     defaultMobileTitle: PropTypes.string,
     headerAction: PropTypes.element,
-    headerText: PropTypes.element,
+    headerText: PropTypes.node,
     onBackClicked: PropTypes.func,
     setMobileScreen: PropTypes.func,
     showBackButton: PropTypes.bool
